+++ chunk-KZKG6LPH.js
@@ -1,0 +1,51 @@
+import {
+  AuthorisedAdminGuard
+} from "./chunk-SRPXUEKS.js";
+import "./chunk-YR74BDOH.js";
+import "./chunk-XI4ZLZAC.js";
+import "./chunk-JNODQGOE.js";
+import "./chunk-LGCOJQOV.js";
+import "./chunk-5XY3WE42.js";
+import "./chunk-UPJHMA72.js";
+import "./chunk-BQISQ53F.js";
+import "./chunk-HT5GXKXQ.js";
+import "./chunk-TPDHL3PI.js";
+import "./chunk-PKJVK52W.js";
+import "./chunk-Q2BV2GZB.js";
+import "./chunk-7A2HMJBQ.js";
+import "./chunk-U7MJINT7.js";
+import "./chunk-Z45QSLBL.js";
+import "./chunk-KWSTWQNB.js";
+
+// src/app/modules/modules.routes.ts
+var ROUTES = [
+  {
+    path: ":id",
+    loadComponent: () => import("./chunk-U6D2A36L.js").then((m) => m.ModulesComponent),
+    children: [
+      {
+        path: "about",
+        loadComponent: () => import("./chunk-MFS54ROW.js").then((m) => m.ModuleAboutComponent)
+      },
+      {
+        path: "systems",
+        loadComponent: () => import("./chunk-RUIIFIP4.js").then((m) => m.ModuleSystemsComponent)
+      },
+      {
+        path: "extend/:id",
+        loadComponent: () => import("./chunk-MBKAFVKO.js").then((m) => m.ExtensionOutletComponent)
+      },
+      {
+        path: "history",
+        canActivate: [AuthorisedAdminGuard],
+        loadComponent: () => import("./chunk-PDZ56LUN.js").then((m) => m.SettingsHistoryViewComponent)
+      },
+      { path: "**", redirectTo: "about" }
+    ]
+  },
+  { path: "**", redirectTo: "-" }
+];
+export {
+  ROUTES
+};
+//# sourceMappingURL=chunk-KZKG6LPH.js.map