+++ chunk-LYNIPNRA.js
@@ -1,0 +1,37 @@
+import "./chunk-KWSTWQNB.js";
+
+// src/app/domains/domains.routes.ts
+var ROUTES = [
+  {
+    path: ":id",
+    loadComponent: () => import("./chunk-3XM3HBM6.js").then((m) => m.DomainsComponent),
+    children: [
+      {
+        path: "about",
+        loadComponent: () => import("./chunk-GAMXVCJI.js").then((m) => m.DomainAboutComponent)
+      },
+      {
+        path: "applications",
+        loadComponent: () => import("./chunk-B76XB5BM.js").then((m) => m.DomainApplicationsComponent)
+      },
+      {
+        path: "authentication",
+        loadComponent: () => import("./chunk-H4IHZIBD.js").then((m) => m.DomainAuthenticationComponent)
+      },
+      {
+        path: "users",
+        loadComponent: () => import("./chunk-BS6GPEVP.js").then((m) => m.DomainUsersComponent)
+      },
+      {
+        path: "extend/:id",
+        loadComponent: () => import("./chunk-MBKAFVKO.js").then((m) => m.ExtensionOutletComponent)
+      },
+      { path: "**", redirectTo: "about" }
+    ]
+  },
+  { path: "**", redirectTo: "-" }
+];
+export {
+  ROUTES
+};
+//# sourceMappingURL=chunk-LYNIPNRA.js.map