+++ chunk-V3RVVYYY.js
@@ -1,0 +1,29 @@
+import "./chunk-KWSTWQNB.js";
+
+// src/app/groups/groups.routes.ts
+var ROUTES = [
+  {
+    path: ":id",
+    loadComponent: () => import("./chunk-A326P52S.js").then((m) => m.GroupsComponent),
+    children: [
+      {
+        path: "about",
+        loadComponent: () => import("./chunk-IIFSM2KS.js").then((m) => m.GroupAboutComponent)
+      },
+      {
+        path: "users",
+        loadComponent: () => import("./chunk-N2U2KSTM.js").then((m) => m.GroupUsersComponent)
+      },
+      {
+        path: "zones",
+        loadComponent: () => import("./chunk-A7BYWTPH.js").then((m) => m.GroupZonesComponent)
+      },
+      { path: "**", redirectTo: "about" }
+    ]
+  },
+  { path: "**", redirectTo: "-" }
+];
+export {
+  ROUTES
+};
+//# sourceMappingURL=chunk-V3RVVYYY.js.map